refactor: combine typeguards
